requirements.txt
ollama==0.3.2
pyperclip==1.9.0
python-dotenv==1.0.0
selenium==4.24.0